requirements.txt
anndata==0.8.0
brotlipy==0.7.0
certifi==2020.6.20
charset-normalizer==2.0.12
click==8.1.2
cloudpickle==2.0.0
cycler==0.11.0
dask==2022.4.1
docker-pycreds==0.4.0
fonttools==4.32.0
fsspec==2022.3.0
gitdb==4.0.9
GitPython==3.1.27
h5py==3.6.0
joblib==1.1.0
kiwisolver==1.4.2
llvmlite==0.38.0
locket==0.2.1
matplotlib==3.5.1
mkl-fft==1.3.1
mkl-service==2.4.0
natsort==8.1.0
networkx==2.8
numba==0.55.1
numpy==1.21.6
packaging==21.3
pandas==1.4.2
partd==1.2.0
pathtools==0.1.2
patsy==0.5.2
Pillow==9.0.1
promise==2.3
protobuf==3.20.0
psutil==5.9.0
pyarrow==7.0.0
pynndescent==0.5.6
pyparsing==3.0.8
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
scanpy==1.9.1
scikit-learn==1.0.2
scipy==1.8.0
seaborn==0.11.2
sentry-sdk==1.5.10
session-info==1.0.0
setproctitle==1.2.3
shortuuid==1.0.8
sklearn==0.0
smmap==5.0.0
statsmodels==0.13.2
stdlib-list==0.8.0
threadpoolctl==3.1.0
toolz==0.11.2
torch==1.11.0
torchaudio==0.11.0
torchvision==0.12.0
tqdm==4.64.0
umap-learn==0.5.3
urllib3==1.26.9
wandb==0.12.14
